+++ packages/Curves/src/fit/fit-inspection.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import * as DG from 'datagrok-api/dg';
-import {FitSeries, FitFunctionType} from '@datagrok-libraries/statistics/src/fit/new-fit-API';
+import {FitSeries, FitFunctionType} from '@datagrok-libraries/statistics/src/fit/fit-engine';
 import {inspectCurve} from './fit-renderer';
 import * as grok from 'datagrok-api/grok';
 
@@ -24,7 +24,7 @@ export function inspectSeries(series: FitSeries, fitFunctionName: FitFunctionTyp
       logX: true,
       title: seriesName,
     },
-    series: [{...series, fit: undefined, fitFunction: fitFunctionName, clickToToggle: true, droplines: ['IC50'], name: seriesName}],
+    series: [{...series, fitFunction: fitFunctionName, clickToToggle: true, droplines: ['IC50'], name: seriesName}],
   }), false);
   const df = DG.DataFrame.fromColumns([curveCol]);
   df.name = seriesName;